feat: implement full UPSERT pipeline with MariaDB benchmark
Implement the real flush_batch in BatchingWriter:
- Resolves dim keys → physical columns via ColumnRegistry
- Builds multi-row INSERT with guarded ON DUPLICATE KEY UPDATE
  (state guard + monotonic timestamp check)
- Executes in chunks of 1000 rows
- Binds base columns (16) + dim columns dynamically

Fix split_sql_statements to strip inline comments (-- after SQL on
same line) which caused _clp_template CREATE to fail.

Fix DimRegistryEntry.width type: SMALLINT UNSIGNED → u16.

Benchmark results (100K records, 8 connections, MariaDB testcontainer):
  Channel-only:  34K rec/s
  Full DB pipeline: 17.6K rec/s
  DB rows verified: 100,000

Usage:
  cargo run -p metalog-bench-ingestion --release -- --with-db

+++ crates/metalog-schema/src/column_registry.rs
@@ -40,7 +40,7 @@ pub struct DimRegistryEntry {
     pub alias_col: Option<String>,
     #[sqlx(rename = "state")]
     pub status: String,
-    pub width: Option<i32>,
+    pub width: Option<u16>,
 }
 
 /// Immutable snapshot of the registry for lock-free reads.
